requirements.txt
# NLP and Language Processing
nltk==3.8.1
chatterbot==1.0.8
chatterbot-corpus==1.2.0

# Machine Learning and Deep Learning
scikit-learn==1.3.2
tensorflow==2.15.0
tflearn==0.5.0
torch==2.1.2

# Data Processing and Numerical Computing
numpy==1.26.0
pandas==2.1.3

# File Handling and Data Formats
# (json and pickle are standard libraries)

# API and Environment Management
groq==0.4.2
python-dotenv==1.0.0

# Web Applications
streamlit==1.29.0
